requirements.txt
cmake==3.22.2
dlib==19.23.0
spacy==3.2.3
spacy-alignments==0.8.4
spacy-legacy==3.0.9
spacy-loggers==1.0.1
spacy-transformers==1.1.4
altgraph==0.17.2
backports.zoneinfo==0.2.1
blis==0.7.6
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
cymem==2.0.6
dateparser==1.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
face-recognition==1.3.0
face-recognition-models==0.3.0
Faker==12.1.0
filelock==3.6.0
future==0.18.2
huggingface-hub==0.4.0
idna==3.3
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.1.0
murmurhash==1.0.6
mypy-extensions==0.4.3
nltk==3.6.7
numpy==1.22.2
packaging==21.3
pathspec==0.9.0
pathy==0.6.1
pefile==2021.9.3
phonenumbers==8.12.42
Pillow==9.0.1
platformdirs==2.5.1
preshed==3.0.6
pydantic==1.8.2
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.2
pyparsing==3.0.7
PySimpleGUI==4.56.0
python-dateutil==2.8.2
python-stdnum==1.17
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
scrubadub==2.0.0
scrubadub-spacy==2.0.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
srsly==2.4.2
text-unidecode==1.3
textblob==0.15.3
thinc==8.0.13
threadpoolctl==3.1.0
tokenizers==0.10.3
tomli==2.0.1
torch==1.11.0
tqdm==4.62.3
transformers==4.15.0
typer==0.4.0
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
wasabi==0.9.0
